Moved structs around
